Revise configure functions
