+++ lib/dataprocessor/pipes/configure.py
@@ -2,88 +2,11 @@
 
 import sys
 import os.path as op
-import yaml
-import ConfigParser as cp
 
 from .. import pipe
-from ..utility import read_configure, check_file
+from ..utility import read_configure, abspath, check_file
 from ..exception import DataProcessorError as dpError
-from ..filetype import FileType
-from ..filetype import guess_filetype_from_path as guess_ft
-
-
-def parse_ini(confpath, section):
-    """
-    Parse .ini and .conf to dictionary
-
-    Parameters
-    ----------
-    confpath : str
-        Path to config file.
-    section : str
-        Specify section name in configure file.
-
-    Returns
-    -------
-    Specified section as a dictionary.
-    """
-    conf = cp.SafeConfigParser()
-    conf.optionxform = str
-    try:
-        read_conf = conf.read(confpath)
-    except cp.MissingSectionHeaderError:
-        raise dpError("Invalid INI file: " + confpath)
-    if not read_conf:
-        raise dpError("Cannot read INI configure file: " + confpath)
-    try:
-        return dict(conf.items(section))
-    except cp.NoSectionError:
-        raise dpError("Section does not found: " + confpath)
-
-
-def parse_yaml(confpath, section):
-    """
-    Parse .yaml to dictionary
-
-    Parameters
-    ----------
-    confpath : str
-        Path to config file.
-    section : str
-        Specify section name in configure file.
-
-    Returns
-    -------
-    Specified section as a dictionary.
-    """
-    with open(confpath, "r") as f:
-        try:
-            d = yaml.load(f)
-        except yaml.YAMLError:
-            raise dpError("Fail to parse YAML file : " + confpath)
-    if section not in d:
-        raise dpError("No such section '{}' in {}".format(section, confpath))
-    return d[section]
-
-
-def get_parser(filetype):
-    """
-    Get parser corresponding to the filetype
-    Parameters
-    ----------
-    filetype : FileType
-        see enum filetype.FileType
-    Returns
-    -------
-    function that takes 2 args, confpath and section.
-    """
-    # check extension in case insensitive way
-    if filetype == FileType.ini:
-        return parse_ini
-    elif filetype == FileType.yaml:
-        return parse_yaml
-    else:
-        return None
+from ..configure import FileType, get_parser, guess_filetype_from_path
 
 
 @pipe.type("run")
@@ -107,18 +30,19 @@ def load(node, filename, filetype=None, section="parameters"):
     """
     NODE_KEY = "configure"
 
-    confpath = check_file(op.join(node["path"], filename))
+    confpath = abspath(op.join(node["path"], filename))
+    check_file(confpath)
     ft = FileType.NONE
 
     if filetype:
         try:
-            ft = FileType[filetype.lower()] 
+            ft = FileType[filetype.lower()]
         except KeyError:
             print >>sys.stderr, "Invalid filetype : " + filetype
             print >>sys.stderr, "Guess from extention"
-            ft = guess_ft(confpath)
+            ft = guess_filetype_from_path(confpath)
     else:
-        ft = guess_ft(confpath)
+        ft = guess_filetype_from_path(confpath)
 
     print ft
     # Invalid filetype
